+++ components/src/main/java/org/patternfly/component/skeleton/Skeleton.java
@@ -23,7 +23,6 @@ import elemental2.dom.HTMLElement;
 
 import static org.jboss.elemento.Elements.div;
 import static org.jboss.elemento.Elements.span;
-import static org.patternfly.core.Ouia.ouiaSafe;
 import static org.patternfly.style.Classes.component;
 import static org.patternfly.style.Classes.modifier;
 import static org.patternfly.style.Classes.screenReader;
@@ -51,7 +50,7 @@ public class Skeleton extends BaseComponent<HTMLElement, Skeleton> {
 
     Skeleton() {
         super(ComponentType.Skeleton, div().css(component(skeleton)).element());
-        ouiaSafe(element(), false);
+        ouiaSafe(false);
     }
 
     // ------------------------------------------------------ builder